+++ src/Pim/Bundle/TransformBundle/Normalizer/Flat/MetricNormalizer.php
@@ -30,19 +30,20 @@ class MetricNormalizer extends AbstractProductValueDataNormalizer
     public function normalize($object, $format = null, array $context = array())
     {
         $context = $this->resolveContext($context);
+        $decimalsAllowed = !array_key_exists('decimals_allowed', $context) || true === $context['decimals_allowed'];
 
         if ('multiple_fields' === $context['metric_format']) {
             $fieldKey = $this->getFieldName($object, $context);
             $unitFieldKey = sprintf('%s-unit', $fieldKey);
 
-            $data = $this->getMetricData($object, false);
+            $data = $this->getMetricData($object, false, $decimalsAllowed);
             $result = [
                 $fieldKey => $data,
                 $unitFieldKey => '' === $data ? '' : $object->getUnit(),
             ];
         } else {
             $result = [
-                $this->getFieldName($object, $context) => $this->getMetricData($object, true),
+                $this->getFieldName($object, $context) => $this->getMetricData($object, true, $decimalsAllowed),
             ];
         }
 
@@ -61,20 +62,22 @@ class MetricNormalizer extends AbstractProductValueDataNormalizer
      *
      * @param MetricInterface $metric
      * @param boolean         $withUnit
+     * @param boolean         $decimalsAllowed
      *
      * @return string
      */
-    public function getMetricData(MetricInterface $metric, $withUnit)
+    public function getMetricData(MetricInterface $metric, $withUnit, $decimalsAllowed = true)
     {
         $data = $metric->getData();
         if (null === $data || '' === $data) {
             return '';
         }
 
+        $pattern = $decimalsAllowed ? '%.4F' : '%d';
         if ($withUnit) {
-            $data = sprintf('%.4F %s', $metric->getData(), $metric->getUnit());
+            $data = sprintf($pattern. ' %s', $metric->getData(), $metric->getUnit());
         } else {
-            $data = sprintf('%.4F', $metric->getData());
+            $data = sprintf($pattern, $metric->getData());
         }
 
         return $data;